+++ src/main/java/de/samply/app/ProjectManagerConst.java
@@ -252,6 +252,7 @@ public class ProjectManagerConst {
     public final static String MANAGE_TOKENS_CRON_EXPRESSION = "MANAGE_TOKENS_CRON_EXPRESSION";
     public final static String CHECK_EXPIRED_ACTIVE_PROJECTS_CRON_EXPRESSION = "CHECK_EXPIRED_ACTIVE_PROJECTS_CRON_EXPRESSION";
     public final static String EXPLORER_URL = "EXPLORER_URL";
+    public final static String ENABLE_TOKEN_MANAGER = "ENABLE_TOKEN_MANAGER";
 
     // Spring Values (SV)
     public final static String HEAD_SV = "${";
@@ -303,12 +304,12 @@ public class ProjectManagerConst {
     public final static String FOCUS_URL_SV = HEAD_SV + FOCUS_URL + BOTTOM_SV;
     public final static String FOCUS_API_KEY_SV = HEAD_SV + FOCUS_API_KEY + BOTTOM_SV;
     public final static String ENABLE_EMAILS_SV = HEAD_SV + ENABLE_EMAILS + ":true" + BOTTOM_SV;
+    public final static String ENABLE_TOKEN_MANAGER_SV = HEAD_SV + ENABLE_TOKEN_MANAGER + ":true" + BOTTOM_SV;
     public final static String MANAGE_TOKENS_CRON_EXPRESSION_SV =
             HEAD_SV + MANAGE_TOKENS_CRON_EXPRESSION + ":#{'0 0 * * * *'}" + BOTTOM_SV;
     public final static String CHECK_EXPIRED_ACTIVE_PROJECTS_CRON_EXPRESSION_SV =
             HEAD_SV + CHECK_EXPIRED_ACTIVE_PROJECTS_CRON_EXPRESSION + ":#{'0 0 1,13 * * *'}" + BOTTOM_SV;
     public final static String EXPLORER_URL_SV = HEAD_SV + EXPLORER_URL + BOTTOM_SV;
-
 
     // Others
     public final static String TEST_EMAIL = "test@project-manager.com";