+++ src/main/java/de/samply/token/dto/DataShieldProjectStatus.java
@@ -3,5 +3,6 @@ package de.samply.token.dto;
 public enum DataShieldProjectStatus {
     CREATED,
     WITH_DATA,
-    NOT_FOUND
+    NOT_FOUND,
+    INACTIVE
 }