+++ src/main/java/de/samply/token/dto/DataShieldTokenStatus.java
@@ -3,5 +3,6 @@ package de.samply.token.dto;
 public enum DataShieldTokenStatus {
     CREATED,
     EXPIRED,
-    NOT_FOUND
+    NOT_FOUND,
+    INACTIVE
 }