+++ CHANGELOG.md
@@ -103,3 +103,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Fetch project users
 - rejectProject state constraints
 - Filter output formats for DataSHIELD
+- Flag ENABLE_TOKEN_MANAGER