+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.0.1 - 2024-02-22]
+## [0.0.1 - 2024-02-26]
 ### Added
 - First version of the project
 - Spring Application
@@ -107,3 +107,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Content Disposition in exposed headers
 - Exists Votum and Application Form
 - Exists Script
+- Explorer Redirect URI